Modified - move waste screen to waste feature
