+++ lib/features/home_feature/presentation/home_screen.dart
@@ -9,7 +9,7 @@ import '../../../core/theme/app_theme.dart';
 import '../../customer_feature/presentation/providers/auth.dart';
 import '../../../screens/article_screen.dart';
 import '../../../screens/collect_list_screen.dart';
-import '../../../screens/waste_cart_screen.dart';
+import '../../waste_feature/presentation/waste_cart_screen.dart';
 import '../../../core/widgets/custom_dialog.dart';
 import '../../store_feature/presentation/screens/product_screen.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
