+++ lib/features/waste_feature/presentation/waste_request_send_screen.dart
@@ -12,19 +12,19 @@ import 'package:tamizshahr/features/waste_feature/business/entities/request_wast
 import 'package:tamizshahr/features/waste_feature/business/entities/wasteCart.dart';
 import 'package:tamizshahr/core/widgets/buton_bottom.dart';
 
-import '../models/customer.dart';
-import '../models/region.dart';
-import '../features/waste_feature/business/entities/address.dart';
+import '../../../models/customer.dart';
+import '../../../models/region.dart';
+import '../business/entities/address.dart';
 
-import '../core/theme/app_theme.dart';
-import '../features/customer_feature/presentation/providers/auth.dart';
-import '../provider/wastes.dart';
-import '../widgets/custom_dialog_enter.dart';
-import '../features/customer_feature/presentation/widgets/custom_dialog_profile.dart';
-import '../widgets/custom_dialog_send_request.dart';
-import '../core/logic/en_to_ar_number_convertor.dart';
-import '../core/widgets/main_drawer.dart';
-import '../core/screens/navigation_bottom_screen.dart';
+import '../../../core/theme/app_theme.dart';
+import '../../customer_feature/presentation/providers/auth.dart';
+import '../../../provider/wastes.dart';
+import '../../../widgets/custom_dialog_enter.dart';
+import '../../customer_feature/presentation/widgets/custom_dialog_profile.dart';
+import '../../../widgets/custom_dialog_send_request.dart';
+import '../../../core/logic/en_to_ar_number_convertor.dart';
+import '../../../core/widgets/main_drawer.dart';
+import '../../../core/screens/navigation_bottom_screen.dart';
 
 class WasteRequestSendScreen extends StatefulWidget {
   static const routeName = '/waste_request_send_screen';
