+++ lib/main.dart
@@ -11,14 +11,14 @@ import 'package:tamizshahr/screens/collect_detail_screen.dart';
 import 'package:tamizshahr/screens/donation_screen.dart';
 import 'package:tamizshahr/features/store_feature/presentation/screens/orders_screen.dart';
 import 'package:tamizshahr/screens/wallet_screen.dart';
-import 'package:tamizshahr/screens/wastes_screen_animated_list.dart';
+import 'package:tamizshahr/features/waste_feature/presentation/wastes_screen_animated_list.dart';
 
 import './provider/articles.dart';
 import 'features/customer_feature/presentation/providers/auth.dart';
 import './provider/messages.dart';
 import './provider/wastes.dart';
 import './screens/about_us_screen.dart';
-import './screens/address_screen.dart';
+import 'features/waste_feature/presentation/address_screen.dart';
 import './screens/article_detail_screen.dart';
 import './screens/article_screen.dart';
 import 'features/store_feature/presentation/screens/cart_screen.dart';
@@ -31,9 +31,9 @@ import './screens/map_screen.dart';
 import './screens/messages_create_screen.dart';
 import 'core/screens/navigation_bottom_screen.dart';
 import 'features/store_feature/presentation/screens/order_products_send_screen.dart';
-import './screens/waste_cart_screen.dart';
-import './screens/waste_request_date_screen.dart';
-import './screens/waste_request_send_screen.dart';
+import 'features/waste_feature/presentation/waste_cart_screen.dart';
+import 'features/waste_feature/presentation/waste_request_date_screen.dart';
+import 'features/waste_feature/presentation/waste_request_send_screen.dart';
 import 'features/waste_feature/presentation/wastes_screen.dart';
 import 'core/constants/strings.dart';
 import 'provider/Products.dart';
